[rocJPEG] Fix uint32 overflow in sample SaveImage buffer sizing
SaveImage() computed per-channel sizes as uint32_t pitch * height and summed
them into a uint32_t before allocating the host buffer. pitch and height
derive from the JPEG SOF marker (up to 65535 each), so large images produce
per-channel sizes exceeding 32 bits. The uint32_t accumulator wraps, yielding
an undersized host buffer while hipMemcpyDtoH copies the full (unwrapped)
device-surface size and the row-by-row fwrite loop reads past the allocation
-- a heap overflow / OOB read reachable from the documented
`rocjpegsample -i evil.jpg` flow.

Perform the size arithmetic in size_t so the products (bounded well within
64 bits) no longer wrap and the buffer is sized correctly.

Fixed in three locations:
- projects/rocjpeg/samples/rocjpeg_samples_utils.h
  (multiplies pitch × aligned_height, overflow at ~65535×65535 aligned)
- projects/rocprofiler-systems/examples/jpegdecode/rocjpeg_samples_utils.h
  (multiplies pitch × height, overflow at 65535×65535)
- projects/rocprofiler-sdk/tests/bin/rocjpeg/rocjpeg_samples_utils.h
  (multiplies pitch × height, overflow at 65535×65535)

All three are vendored copies of the same utility in different subprojects.
They share identical vulnerable logic but have diverged enough (different
dependencies, formatting) that symlinking would require a broader refactor.

Added regression test:
- projects/rocjpeg/test/rocjpeg_negative_api_tests/test_saveimage_overflow.cpp
  Verifies that large JPEG dimensions (approaching 65535×65535) do not cause
  integer overflow when computing the host buffer size. The test exercises
  the size_t arithmetic path and confirms the computed buffer size exceeds
  UINT32_MAX for max-dimension images (which would have wrapped to ~0 with
  the old uint32_t accumulator).

Fixes: SWSPLAT-24294

+++ projects/rocjpeg/samples/rocjpeg_samples_utils.h
@@ -571,11 +571,15 @@ public:
         aliged_heights[0] = align(heights[0], mem_alignment);
         aliged_heights[1] = align(heights[1], mem_alignment);
         aliged_heights[2] = align(heights[2], mem_alignment);
-        uint32_t channel0_size = output_image->pitch[0] * aliged_heights[0];
-        uint32_t channel1_size = output_image->pitch[1] * aliged_heights[1];
-        uint32_t channel2_size = output_image->pitch[2] * aliged_heights[2];
+        // Compute per-channel sizes in size_t. pitch and aligned height each derive
+        // from the JPEG dimensions (up to 65535, then alignment-padded), so their
+        // product can exceed 32 bits. Accumulating in uint32_t would wrap and yield an
+        // undersized host buffer while hipMemcpyDtoH still copies the full surface.
+        size_t channel0_size = static_cast<size_t>(output_image->pitch[0]) * aliged_heights[0];
+        size_t channel1_size = static_cast<size_t>(output_image->pitch[1]) * aliged_heights[1];
+        size_t channel2_size = static_cast<size_t>(output_image->pitch[2]) * aliged_heights[2];
 
-        uint32_t output_image_size = channel0_size + channel1_size + channel2_size;
+        size_t output_image_size = channel0_size + channel1_size + channel2_size;
 
         if (hst_ptr == nullptr) {
             hst_ptr = new uint8_t [output_image_size];
